backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiocache>=0.12.3",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.1",
    "aiosignal>=1.4.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.17.0",
    "annotated-types>=0.7.0",
    "anyio>=4.11.0",
    "asyncpg>=0.30.0",
    "attrs>=25.4.0",
    "bcrypt>=5.0.0",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "dnspython>=2.8.0",
    "ecdsa>=0.19.1",
    "email-validator>=2.3.0",
    "fastapi>=0.119.1",
    "frozenlist>=1.8.0",
    "greenlet>=3.2.4",
    "h11>=0.16.0",
    "httpx>=0.28.1",
    "idna>=3.11",
    "iniconfig>=2.3.0",
    "jinja2>=3.1.6",
    "mako>=1.3.10",
    "markupsafe>=3.0.3",
    "multidict>=6.7.0",
    "packaging>=25.0",
    "pluggy>=1.6.0",
    "propcache>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pyasn1>=0.6.1",
    "pydantic>=2.12.3",
    "pydantic-core>=2.41.4",
    "pydantic-settings>=2.11.0",
    "pygments>=2.19.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-json-logger>=2.0.7",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=7.1.0",
    "rsa>=4.9.1",
    "six>=1.17.0",
    "slowapi>=0.1.9",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.44",
    "starlette>=0.48.0",
    "structlog>=24.1.0",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "uvicorn>=0.38.0",
    "yarl>=1.22.0",
]
